Refactor logic around subteam, it is now only specified on team level
Having it as property of a sprint and as a property of team lead to a lot of unnecessary complexity. However, identifying a sprint only based on ID proved troublesome, since sprint ID is not unique when using subteams. Solved by creating a unique identifier for sprints based on sprint id and board id.